imagenes en about,cambios en perfil

+++ frontend/src/pages/Perfil.jsx
@@ -6,13 +6,15 @@ import { UserContext } from '../context/UserContext.jsx';
 import Footer from '../components/Footer.jsx';
 import Nav from '../components/Nav.jsx';
 import '../styles/Perfil.css';
+import { Tab, Tabs, TabList, TabPanel } from 'react-tabs';
+import 'react-tabs/style/react-tabs.css';
+import Swal from 'sweetalert2';
 
 function Perfil() {
   const { user, handleLogout } = useContext(UserContext);
   const navigate = useNavigate();
   const { id } = useParams();
 
-  // Estado inicializado con valores vacíos por defecto
   const [userUpdate, setUserUpdate] = useState({
     phone: user ? user.telefono : '',
     email: user ? user.email : '',
@@ -22,8 +24,6 @@ function Perfil() {
   const [purchaseHistory, setPurchaseHistory] = useState([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
-  const [tableVisible, setTableVisible] = useState(false); // Estado para controlar la visibilidad de la tabla
-
   const [showForm, setShowForm] = useState(false);
 
   useEffect(() => {
@@ -61,9 +61,26 @@ function Perfil() {
   };
 
   const handleDeleteProfile = () => {
-    deleteUser(user);
-    handleLogout();
-    navigate('/register');
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: "Si elimina su cuenta perdera sus datos y tendrá que volver a registrarse",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#e01717',
+      cancelButtonColor: '#126606',
+      confirmButtonText: 'Sí, eliminarla'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        deleteUser(user);
+        handleLogout();
+        navigate('/register');
+        Swal.fire(
+          '¡Eliminada!',
+          'Tu cuenta ha sido eliminada.',
+          'success'
+        )
+      }
+    });
   };
 
   if (!user) {
@@ -74,88 +91,87 @@ function Perfil() {
     <>
       <div className='container-perfil'>
         <Nav />
-        <div className='container-'>
-          <h2>{user.name}</h2>
-          <p>{user.email}</p>
-        </div>
+         <Tabs className='content'>
+          <TabList>
+            <Tab>Perfil</Tab>
+            <Tab>Historial de Compras</Tab>
+          </TabList>
 
-        <div className="container-form">
-         
-            <button onClick={() => setShowForm(!showForm)}>
-              {showForm ? 'Ocultar perfil' :'Mostrar perfil'}
-            </button>
-          
-          {showForm && ( // Solo muestra el formulario si showForm es true
-            <form className="form" onSubmit={handleSubmit}>
-              <label htmlFor='email'>Email</label>
-              <input type='text' name='email' id='email' value={userUpdate.email || ''} onChange={handleChange} />
-              <label htmlFor='phone'>Telefono</label>
-              <input type='phone' name='phone' id='phone' value={userUpdate.phone || ''} onChange={handleChange} />
-              <label htmlFor='password'>Contraseña</label>
-              <input type='password' name='password' id='password' value={userUpdate.password || ''} onChange={handleChange} />
-              <button className='btn-guardar' type='submit'>Guardar</button>
-              <button className='btn-delete' type='button' onClick={handleDeleteProfile}>Eliminar perfil</button>
-            </form>
-          )}
-        </div>
-
-        <div className="container-purchase-history">
-
-
-          <button onClick={() => setTableVisible(!tableVisible)}> {/* Botón para mostrar/ocultar la tabla */}
-            {tableVisible ? 'Ocultar historial de compras' : 'Mostrar historial de compras'}
-          </button>
-
-          {loading && <p>Cargando...</p>}
-          {error && <p>Error: {error}</p>}
-          {tableVisible && (
-            <div className='puserchase'>
-              <table>
-                <thead>
-                  <tr>
-                    <th>Fecha de Factura</th>
-                    <th>Total Sin IVA</th>
-                    <th>Total Con IVA</th>
-                    <th>Producto</th>
-                    <th>Descripción</th>
-                    <th>Precio</th>
-                    <th>Cantidad</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {purchaseHistory.map((invoice) => (
-                    <tr key={invoice.idInvoice}>
-                      <td>{new Date(invoice.invoiceDate).toLocaleDateString()}</td>
-                      <td>${invoice.total_without_iva}</td>
-                      <td>${invoice.total_with_iva}</td>
-                      <td>
-                        {invoice.invoiceDetails.map((detail) => (
-                          <p key={detail.id}>{detail.product.product}</p>
-                        ))}
-                      </td>
-                      <td>
-                        {invoice.invoiceDetails.map((detail) => (
-                          <p key={detail.id}>{detail.product.description}</p>
-                        ))}
-                      </td>
-                      <td>
-                        {invoice.invoiceDetails.map((detail) => (
-                          <p key={detail.id}>${detail.product.price}</p>
-                        ))}
-                      </td>
-                      <td>
-                        {invoice.invoiceDetails.map((detail) => (
-                          <p key={detail.id}>{detail.amount_sold}</p>
-                        ))}
-                      </td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
+          <TabPanel>
+            <div className='user-profile'>
+              <h2>Bienvenido/a {user.name}</h2>
+              <p>Su cuenta: {user.email}</p>
+              <button onClick={() => setShowForm(!showForm)}>
+                {showForm ? 'Ocultar perfil' : 'Mostrar perfil'
+                }
+              </button>
+              {showForm && (
+                <form className="form" onSubmit={handleSubmit}>
+                  <label htmlFor='email'>Email</label>
+                  <input type='text' name='email' id='email' value={userUpdate.email || ''} onChange={handleChange} />
+                  <label htmlFor='phone'>Teléfono</label>
+                  <input type='phone' name='phone' id='phone' value={userUpdate.phone || ''} onChange={handleChange} />
+                  <label htmlFor='password'>Contraseña</label>
+                  <input type='password' name='password' id='password' value={userUpdate.password || ''} onChange={handleChange} />
+                  <button className='btn-guardar' type='submit'>Guardar</button>
+                  <button className='btn-delete' type='button' onClick={handleDeleteProfile}>Eliminar perfil</button>
+                </form>
+              )}
             </div>
-          )}
-        </div>
+          </TabPanel>
 
+          <TabPanel>
+            <div className='purchase-history'>
+              <h3>Historial de Compras</h3>
+              {loading && <p>Cargando...</p>}
+              {error && <p>Aún no tiene compras realizadas</p>}
+              <div className='puserchase'>
+                <table className="table">
+                  <thead>
+                    <tr>
+                      <th>Fecha de Factura</th>
+                      <th>Total Sin IVA</th>
+                      <th>Total Con IVA</th>
+                      <th>Producto</th>
+                      <th>Descripción</th>
+                      <th>Precio</th>
+                      <th>Cantidad</th>
+                    </tr>
+                  </thead>
+                  <tbody>
+                    {purchaseHistory.map((invoice) => (
+                      <tr key={invoice.idInvoice}>
+                        <td>{new Date(invoice.invoiceDate).toLocaleDateString()}</td>
+                        <td>${invoice.total_without_iva}</td>
+                        <td>${invoice.total_with_iva}</td>
+                        <td>
+                          {invoice.invoiceDetails.map((detail) => (
+                            <p key={detail.id}>{detail.product.product}</p>
+                          ))}
+                        </td>
+                        <td>
+                          {invoice.invoiceDetails.map((detail) => (
+                            <p key={detail.id}>{detail.product.description}</p>
+                          ))}
+                        </td>
+                        <td>
+                          {invoice.invoiceDetails.map((detail) => (
+                            <p key={detail.id}>${detail.product.price}</p>
+                          ))}
+                        </td>
+                        <td>
+                          {invoice.invoiceDetails.map((detail) => (
+                            <p key={detail.id}>{detail.amount_sold}</p>
+                          ))}
+                        </td>
+                      </tr>
+                    ))}
+                  </tbody>
+                </table>
+              </div>
+            </div>
+          </TabPanel>
+        </Tabs>
         <Footer />
       </div>
     </>
